Move component specs into their component folders
This moves the individual component specs to live with their application code. The specs have also been renamed to better match the component conventions.
