+++ lib/tasks/test.js
@@ -12,7 +12,10 @@ gulp.task('test:lib', () => gulp.src(paths.testSpecs + 'transpiler_spec.js', {re
   .pipe(mocha({reporter: 'spec'}))
 )
 
-gulp.task('test:components', () => gulp.src(paths.testSpecs + 'components' + '**/*', {read: false})
+gulp.task('test:components', () => gulp.src([
+  paths.testSpecs + 'components/' + '**/*',
+  paths.srcComponents + '**/*.spec.js'
+], {read: false})
   .pipe(mocha({reporter: 'spec'}))
 )
