+++ src/components/form-radio-group/form-radio-group.spec.js
@@ -1,4 +1,4 @@
-const expectComponent = require('./helper').expectComponent
+const expectComponent = require('../../../test/specs/components/helper').expectComponent
 
 describe('Form radio group component', () => {
   it('should render', () => {
